Bugfix for having more than MAX_PETS in Party
Having more than MAX_PETS, defined in
config.json, in the party truncates the party
display in the freezer, not allowing the "excess"
pets to be shown or selected.

Bugfix handles "excess" pets gracefully:

Change the game so it displays the entire party
in the freezer menu, and any pet can be selected.

Does not change MAX_PETS, so putting "excess"
pets in the freezer does not allow adding pets to
the party until the size of the party is less
than MAX_PETS.